style(layout): change loader design color (#47)
### TL;DR

The color of the NextTopLoader has been updated from `#2acf80` to `#f9792f`.

### What changed?

The color code in the `NextTopLoader` definition in the Root Layout component file (`layout.tsx`) was changed to update the color scheme of the loader.

### How to test?

To view the new loader color, just start the application normally. The NextTopLoader appears at the top of the page during route changes.

### Why make this change?

The color change was made to match the updated color scheme of the website.

---